+++ CHANGELOG.md
@@ -6,8 +6,9 @@
 
 - add UI for "format" parameter in "todate" step (not supported for mongo36
   translator)
-- expose `setCodeEditor` to set a custom code editor component for  custom steps
+- expose `setCodeEditor` to set a custom code editor component for custom steps
 - custom step which allows users to edit manually a step
+- uniquegroups step to get unique groups/values" from columns
 
 ## [0.10.0] - 2020-02-06