+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## UNRELEASED -
+## [0.11.0] - 2020-02-12
 
 ### Added
 
@@ -157,6 +157,7 @@
 
 - Initial version, showtime!
 
+[0.11.0]: https://github.com/ToucanToco/weaverbird/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/ToucanToco/weaverbird/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/ToucanToco/weaverbird/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/ToucanToco/weaverbird/compare/v0.7.0...v0.8.0